Resample filter fix

+++ sources/core/sample/sampleutils.h
@@ -32,7 +32,7 @@ class SampleUtils
 public:
     static QVector<float> int24ToFloat(const qint16 * data16, const quint8 * data24, quint32 length);
     static void floatToInt24(const QVector<float> data, qint16 *& data16, quint8 *& data24);
-    static QVector<float> resampleMono(QVector<float> vData, double echInit, quint32 echFinal);
+    static QVector<float> resampleMono(QVector<float> vData, double echInit, double echFinal);
     static void fillSincTable(float * table, int order, int subdivisions, double kaiserBesserAlpha);
     static QVector<float> bandFilter(QVector<float> vData, double dwSmplRate, double fBas, double fHaut, int ordre);
     static QVector<float> cutFilter(QVector<float> vData, quint32 dwSmplRate, QVector<float> dValues, int maxFreq);
